Sheol: Remove passive blue slime attack.
Blue slimes have an active and a passive attack but the passive attack
apparently does not work. However, instead of making it work, I think it is
better that they don't have the attack, like it is now. This commit prevents
the passive attack from showing up in any (up-to-date) external collections
monster information (such as the Pinobot and the Nethackwiki).



git-svn-id: http://svn.code.sf.net/p/unnethack/code/trunk@1555 f1e491ac-5ded-4967-80bf-9f4855be876a