Make rustproof/erodeproof/fixed known by default (from DynaHack).
All cases of rustproofing in the game are known, so having the rknown
bit off by default only meant cluttering item menus during ID with
entries which had everything known about them except this one thing.

This also causes rustproofing to be known when reciting a scroll of
enchant weapon/armor while blind and confused, since the result is
unambiguous.

git-svn-id: http://svn.code.sf.net/p/unnethack/code/trunk@1798 f1e491ac-5ded-4967-80bf-9f4855be876a